+++ core/src/main/java/com/epam/esm/service/CertificatesService.java
@@ -1,7 +1,6 @@
 package com.epam.esm.service;
 
 import com.epam.esm.dto.CertificateDTO;
-import com.epam.esm.model.Certificate;
 
 import java.util.List;
 
@@ -9,11 +8,11 @@ public interface CertificatesService {
 
     List<CertificateDTO> readAll(String tagName, String partName, String partDescription, String dateSort, String nameSort);
 
-    CertificateDTO read(int id);
+    CertificateDTO read(long id);
 
     CertificateDTO create(CertificateDTO certificateDTO);
 
-    CertificateDTO update(int id, CertificateDTO certificateDTO);
+    CertificateDTO update(long id, CertificateDTO certificateDTO);
 
-    CertificateDTO delete(int id);
+    CertificateDTO delete(long id);
 }
